de-duplicate condition fixture definition
